tailwindify Icon and move to primitives
