--- src/components/Icon.css
@@ -1,3 +1,0 @@
-.Icon_flipHorizontally > svg {
-  transform: scale(-1, 1);
-}
