+++ src/components/RouteLeg.jsx
@@ -3,7 +3,7 @@ import { useIntl } from 'react-intl';
 import { DEFAULT_PT_COLOR, getTextColor } from '../lib/colors';
 import ModeIcon from './ModeIcon';
 import { MODES } from '../lib/TransitModes';
-import Icon from './Icon';
+import Icon from './primitives/Icon';
 import { ReactComponent as Bicycle } from 'iconoir/icons/bicycle.svg';
 import { ReactComponent as WarningTriangle } from 'iconoir/icons/warning-triangle.svg';
 import { formatInterval } from '../lib/time';
