+++ src/components/SelectionListItem.jsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { useIntl } from 'react-intl';
 import classnames from 'classnames';
-import Icon from './Icon';
+import Icon from './primitives/Icon';
 
 import { ReactComponent as CancelIcon } from 'iconoir/icons/cancel.svg';
